+++ store/modules/data/enums.js
@@ -1,8 +1,8 @@
 export const TWITCH = {
   METRIC_IDS: {
+    "Twitch Frequency": "472d0707-ff87-4198-9374-c28900bb216c",
     "Twitch Force": "89cf1105-a015-434f-b527-4169b9400e26",
     "Twitch Period": "6e0cd81c-7861-4c49-ba14-87b2739d65fb",
-    "Twitch Frequency": "472d0707-ff87-4198-9374-c28900bb216c",
     "Twitch Width 80": "400ff0fb-a29e-446f-a4d5-11b785f4c69f",
     "Contraction Velocity": "73961e7c-17ec-42b0-b503-a23195ec249c",
     "Relaxation Velocity": "0fcc0dc3-f9aa-4f1b-91b3-e5b5924279a9",