+++ store/modules/gradient/mutations.js
@@ -14,6 +14,6 @@ export default {
   },
   reset_gradient_range(state) {
     state.gradient_range_min = 0;
-    state.gradient_range_max = 100;
+    state.gradient_range_max = 1;
   },
 };