+++ store/modules/heatmap/enums.js
@@ -1,7 +1,7 @@
 export const METRIC_UNITS = {
+  "Twitch Frequency": "Hz",
   "Twitch Force": "µN",
   "Twitch Period": "s",
-  "Twitch Frequency": "Hz",
   "Twitch Width 80": "--",
   "Contraction Velocity": "--",
   "Relaxation Velocity": "--",